Few names changed (LOOK on description below), added toasts where needed, modified look of NewMeetingActivity MainMenuActivity -> AudioOnTouchActivity AudioOnTouchActivity -> MainMenuActivity These names needed swap, New -> NewMeetingActivity
When input meeting name there is only hint now, when name is empty only toast
is showed and app waits for further actions.
When no emails, toast is shown (in order to exception) that there is no one
to who we can send mail.

Few variables changed like pliki -> files